+++ src/http-client/echo-server/Cargo.toml
@@ -23,5 +23,5 @@ publish           = false
 axum       = "0.8.1"
 serde_json = "1"
 tokio      = { version = "1.42", features = ["macros"] }
-gax        = { path = "..", package = "google-cloud-gax" }
+gax        = { path = "../../gax", package = "google-cloud-gax" }
 rpc        = { path = "../../generated/rpc/types", package = "google-cloud-rpc" }
